backend/requirements.txt
PyGithub
numpy
fastapi
uvicorn
pandas